+++ lib/qtermwidget/qtermwidget.cpp
@@ -558,7 +558,12 @@ QString QTermWidget::keyBindings()
 
 void QTermWidget::toggleShowSearchBar()
 {
-    m_searchBar->isHidden() ? m_searchBar->show() : m_searchBar->hide();
+    if(m_searchBar->isHidden()) {
+        m_searchBar->setText(selectedText(true));
+        m_searchBar->show();
+    } else {
+        m_searchBar->hide();
+    }
 }
 
 bool QTermWidget::flowControlEnabled(void)