+++ CHANGELOG.md
@@ -12,6 +12,7 @@ en-US:
 - Add up to four window split screen mode and multiple layout modes
 - Add single/group/all session three modes in the send command window
 - Improve the context menu of URL recognition links
+- Improve the find window to automatically fill in the currently selected text each time it is opened
 - Fix the Telnet session storage configuration error in non-English environments, which causes connection problems [#IAADHZ](https://gitee.com/QQxiaoming/quardCRT/issues/IAADHZ)
 - Add pre-built plugins [timestamp](https://github.com/QuardCRT-platform/plugin-timestamp), and update pre-built plugin [ListSerial](https://github.com/QuardCRT-platform/plugin-ListSerial) to V0.0.3
 
@@ -25,6 +26,7 @@ zh-CN:
 - 增加了至多四窗口分屏模式以及多种布局模式
 - 发送命令窗口增加单会话/组会话/全部会话三种模式
 - 改进URL识别链接的上下文菜单
+- 改进查找窗口在每次打开时自动填入当前选择的文本
 - 修复非英文环境下Telnet会话存储配置错误导致无法连接问题 [#IAADHZ](https://gitee.com/QQxiaoming/quardCRT/issues/IAADHZ)
 - 新增预构建插件[timestamp](https://github.com/QuardCRT-platform/plugin-timestamp)，更新预构建插件[ListSerial](https://github.com/QuardCRT-platform/plugin-ListSerial)到V0.0.3版本